rn: Move the header styles to their component
