+++ packages/edge-login-ui-rn/src/components/common/Header.js
@@ -7,6 +7,9 @@ import LinearGradient from 'react-native-linear-gradient'
 import s from '../../common/locales/strings.js'
 import T from '../../components/common/FormattedText.js'
 import * as Colors from '../../constants/Colors.js'
+import * as Constants from '../../constants/index.js'
+import * as Styles from '../../styles/index.js'
+import { scale } from '../../util/scaling.js'
 import { Button } from '../common/Button.js'
 import { HeaderBackButton } from '../common/HeaderBackButton.js'
 
@@ -16,7 +19,6 @@ type Props = {
   showBackButton: boolean,
   showSkipButton?: boolean,
   skipScreen?: () => void,
-  style: Object,
   subTitle: string,
   title: string,
   useCancel?: boolean
@@ -24,7 +26,7 @@ type Props = {
 
 export class Header extends Component<Props> {
   render() {
-    const Style = this.props.style
+    const Style = HeaderContainerScaledStyle
     return (
       <LinearGradient
         style={[Style.container, { paddingTop: 0 }]}
@@ -39,7 +41,7 @@ export class Header extends Component<Props> {
     )
   }
 
-  renderBack(style: Object) {
+  renderBack(style: typeof HeaderContainerScaledStyle) {
     if (!this.props.showBackButton) {
       return
     }
@@ -92,5 +94,85 @@ export class Header extends Component<Props> {
 
   onBack = () => {
     this.props.goBack()
+  }
+}
+
+const HeaderContainerScaledStyle = {
+  container: {
+    position: 'relative',
+    height: scale(Constants.HEADER_HEIGHT),
+    width: '100%',
+    backgroundColor: Constants.TRANSPARENT,
+    flexDirection: 'row',
+    paddingTop: scale(6)
+  },
+  headerBackButtonStyle: {
+    backButton: {
+      flexDirection: 'row',
+      alignItems: 'center'
+    },
+    backIconStyle: {
+      position: 'relative',
+      top: scale(1),
+      paddingLeft: scale(10),
+      paddingRight: scale(6),
+      color: Constants.WHITE,
+      fontSize: scale(20)
+    },
+    sideText: {
+      color: Constants.WHITE,
+      fontSize: scale(14)
+    },
+    icon: {
+      color: Constants.WHITE
+    },
+    default: {
+      backgroundColor: Constants.TRANSPARENT,
+      color: Constants.WHITE
+    }
+  },
+  left: {
+    flex: 1,
+    justifyContent: 'center'
+  },
+  center: {
+    flex: 3,
+    justifyContent: 'center'
+  },
+  right: {
+    flex: 1,
+    justifyContent: 'center'
+  },
+  headlineText: {
+    fontSize: scale(17),
+    width: '100%',
+    textAlign: 'center',
+    color: Constants.WHITE
+  },
+  subHeadText: {
+    fontSize: scale(11),
+    width: '100%',
+    textAlign: 'center',
+    color: Constants.ACCENT_MINT
+  },
+  textButton: {
+    upStyle: { ...Styles.TextOnlyButtonUpScaledStyle, width: '100%' },
+    upTextStyle: {
+      ...Styles.TextOnlyButtonTextUpScaledStyle,
+      width: '100%',
+      color: Constants.WHITE,
+      fontSize: scale(Constants.FONTS.defaultButtonTextSize)
+    },
+    downTextStyle: {
+      ...Styles.TextOnlyButtonTextDownScaledStyle,
+      width: '100%',
+      color: Constants.SECONDARY,
+      fontSize: scale(Constants.FONTS.defaultButtonTextSize)
+    },
+    downStyle: {
+      ...Styles.TextOnlyButtonDownScaledStyle,
+      width: '100%',
+      height: scale(50)
+    }
   }
 }
