+++ packages/edge-login-ui-rn/src/styles/index.js
@@ -7,7 +7,6 @@ export * from './common/CenteredSpinnerStyle'
 export * from './common/CheckboxStyles'
 export * from './common/Fonts'
 export * from './common/FormFieldStyle'
-export * from './common/HeaderStyles'
 export * from './common/InputStyles'
 export * from './common/ListItemStyles'
 export * from './common/LogoImageStyles'
